refactor: extract add monitoring point component
