+++ client/src/app/dashboard/monitoring/page.tsx
@@ -6,35 +6,14 @@ import Button from '@mui/material/Button';
 import Stack from '@mui/material/Stack';
 import Typography from '@mui/material/Typography';
 import { PlusIcon } from '@phosphor-icons/react/dist/ssr/Plus';
-import {
-  Dialog,
-  DialogContent,
-  DialogTitle,
-  MenuItem,
-  Select,
-  TextField,
-  FormControl,
-  InputLabel,
-} from '@mui/material';
 import { AppDispatch, RootState } from '@/store';
 import { useDispatch, useSelector } from 'react-redux';
-import { addMonitoringPoint, fetchMachines, SensorType } from '@/store/features/machinesSlice';
+import { fetchMachines } from '@/store/features/machinesSlice';
 import { useEffect } from 'react';
 import { MonitoringPointsTable, type MonitoringPoint } from '@/components/dashboard/monitoring/monitoring-points-table';
+import { AddMonitoringPointDialog } from '@/components/dashboard/monitoring/add-monitoring-point-dialog';
 
 type Order = 'asc' | 'desc';
-
-// Sensor type configurations
-const SENSOR_CONFIGS = {
-  pump: [
-    { value: 'HF+', label: 'Pressure/Level/Flow (HF+)' }
-  ],
-  fan: [
-    { value: 'TcAg', label: 'Temperature (TcAg)' },
-    { value: 'TcAS', label: 'Humidity (TcAS)' },
-    { value: 'HF+', label: 'Pressure/Level/Flow (HF+)' }
-  ]
-} as const;
 
 export default function Page(): React.JSX.Element {
   const dispatch = useDispatch<AppDispatch>();
@@ -42,9 +21,6 @@ export default function Page(): React.JSX.Element {
   const [orderBy, setOrderBy] = React.useState<keyof MonitoringPoint>('id');
   const [order, setOrder] = React.useState<Order>('asc');
   const [open, setOpen] = React.useState(false);
-  const [selectedMachineId, setSelectedMachineId] = React.useState<string>('');
-  const [monitoringPointName, setMonitoringPointName] = React.useState<string>('');
-  const [sensorType, setSensorType] = React.useState<string>('');
 
   useEffect(() => {
     dispatch(fetchMachines());
@@ -62,32 +38,9 @@ export default function Page(): React.JSX.Element {
     setOrderBy(property);
   };
 
-  const handleAddMonitoringPoint = () => {
-    dispatch(addMonitoringPoint({ machineId: selectedMachineId, monitoringPoint: { id: '', monitoringPointName, sensorType: sensorType as SensorType } }));
-    // Reset form
-    setSelectedMachineId('');
-    setMonitoringPointName('');
-    setSensorType('');
-    setOpen(false);
-  };
-
   const handleCloseDialog = () => {
-    setSelectedMachineId('');
-    setMonitoringPointName('');
-    setSensorType('');
     setOpen(false);
   };
-
-  const handleMachineChange = (machineId: string) => {
-    setSelectedMachineId(machineId);
-    // Reset sensor type when machine changes
-    setSensorType('');
-  };
-
-  // Get the selected machine to determine available sensor types
-  const selectedMachine = machines.find(machine => machine.id === selectedMachineId);
-  const machineType = selectedMachine?.type as keyof typeof SENSOR_CONFIGS;
-  const availableSensors = machineType ? SENSOR_CONFIGS[machineType] : [];
 
   return (
     <Stack spacing={3}>
@@ -113,63 +66,11 @@ export default function Page(): React.JSX.Element {
         onSort={handleSort}
       />
       
-      <Dialog open={open} onClose={handleCloseDialog}>
-        <DialogTitle>Add Monitoring Point</DialogTitle>
-        <DialogContent>
-          <Stack spacing={3} sx={{ mt: 2, minWidth: 400 }}>
-            <FormControl fullWidth>
-              <InputLabel>Machine Name</InputLabel>
-              <Select
-                value={selectedMachineId}
-                label="Machine Name"
-                onChange={(e) => handleMachineChange(e.target.value as string)}
-              >
-                {machines.map((machine) => (
-                  <MenuItem key={machine.id} value={machine.id}>
-                    {machine.name} ({machine.type})
-                  </MenuItem>
-                ))}
-              </Select>
-            </FormControl>
-            
-            <TextField
-              fullWidth
-              label="Monitoring Point Name"
-              value={monitoringPointName}
-              onChange={(e) => setMonitoringPointName(e.target.value)}
-            />
-            
-            <FormControl fullWidth>
-              <InputLabel>Sensor Type</InputLabel>
-              <Select
-                value={sensorType}
-                label="Sensor Type"
-                onChange={(e) => setSensorType(e.target.value as string)}
-                disabled={!selectedMachineId}
-              >
-                {availableSensors.map((sensor) => (
-                  <MenuItem key={sensor.value} value={sensor.value}>
-                    {sensor.label}
-                  </MenuItem>
-                ))}
-              </Select>
-            </FormControl>
-            
-            <Stack direction="row" spacing={2} sx={{ mt: 2 }}>
-              <Button 
-                variant="contained" 
-                onClick={handleAddMonitoringPoint}
-                disabled={!selectedMachineId || !monitoringPointName || !sensorType}
-              >
-                Add Monitoring Point
-              </Button>
-              <Button variant="outlined" onClick={handleCloseDialog}>
-                Cancel
-              </Button>
-            </Stack>
-          </Stack>
-        </DialogContent>
-      </Dialog>
+      <AddMonitoringPointDialog
+        open={open}
+        onClose={handleCloseDialog}
+        machines={machines}
+      />
     </Stack>
   );
 }
